% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-prepareData.R
\name{prepareData}
\alias{prepareData}
\title{prepareData}
\usage{
prepareData(
  rawData,
  in_x,
  in_xUnc,
  in_y,
  in_yUnc,
  in_xCategorical,
  in_xCatUnc,
  in_regType
)
}
\arguments{
\item{rawData}{(reactive) data()}

\item{in_x}{(reactive) input$x}

\item{in_xUnc}{(reactive) input$xUnc}

\item{in_y}{(reactive) input$y}

\item{in_yUnc}{(reactive) input$yUnc}

\item{in_xCategorical}{(reactive) input$xCategorical}

\item{in_xCatUnc}{(reactive) input$xCatUnc}

\item{in_regType}{(reactive) input$regType}
}
\description{
prepareData
}
