% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-NAMESPACE.R, R/03-dataInput.R,
%   R/03-modelEstimation.R
\name{shinyModule}
\alias{shinyModule}
\alias{dataInputUI}
\alias{dataInput}
\alias{modelEstimationUI}
\alias{modelEstimation}
\title{Server and UI Functions for Shiny Module}
\usage{
dataInputUI(id, title = "")

dataInput(input, output, session, config)

modelEstimationUI(id, title = "")

modelEstimation(input, output, session, data, config)
}
\arguments{
\item{id}{namespace id}

\item{title}{title of tab in tabset panel}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session}

\item{config}{(list) list of configuration parameters}

\item{data}{data reactive}
}
\description{
Server and UI Functions for Shiny Module
}
