% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUploadModel.R
\name{uploadModelServer}
\alias{uploadModelServer}
\title{Server function upload model}
\usage{
uploadModelServer(id, reset = reactive(FALSE))
}
\arguments{
\item{id}{namespace id}

\item{reset}{(reactive) resets the selection of the online model}
}
\description{
Backend for upload model module
}
