% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUploadModel.R
\name{downloadModelServer}
\alias{downloadModelServer}
\title{Server function download model}
\usage{
downloadModelServer(id, dat, inputs, model)
}
\arguments{
\item{id}{namespace id}

\item{dat}{(reactive) user data}

\item{inputs}{(reactive) user inputs}

\item{model}{(reactive) model output object}
}
\description{
Backend for download model module
}
