% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUploadModel.R
\name{downloadModelServer}
\alias{downloadModelServer}
\title{Server function download model}
\usage{
downloadModelServer(id, inputs, model, uploadedNotes)
}
\arguments{
\item{id}{namespace id}

\item{uploadedNotes}{(reactive) variable that stores content of README.txt}

\item{allParentInput}{(reactive) list of inputs from parent module}

\item{yEstimates}{(reactive) An object created by \code{\link{estimateY}}.
Distributions of the dependent variable.}

\item{formulas}{(reactive) formulas}

\item{data}{(reactive) data}
}
\description{
Backend for download model module
}
