% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelEvaluation.R
\name{getICData}
\alias{getICData}
\title{Get Data of IC}
\usage{
getICData(allFits, modelNames, ic, withColumnICName = FALSE)
}
\arguments{
\item{ic}{(character) name of information criterion, e.g. \code{"AUC", "Rsq", "RsqAdj", "Bayes_Rsq", "df", "logLik", "nagelkerke", "Loo", "WAIC"}}
}
\description{
Get Data of IC
}
